requirements.txt
django
requests
djangorestframework
django-filter